requirements.txt
asgiref==3.6.0
Django==4.1.8
Pillow==9.5.0
sqlparse==0.4.3
